init Medea Destiny's AO Link
Medea Destiny's AO interface script for OpenCollar

This script is inteneded to be a drop-in for ZHAO-2 and Oracul AOs that
are not already compatible with OpenCollar's AO interface.

Scope: The script tries to identify the AO type by checking script
naming. It receives the ZHAO ON, ZHAO OFF and ZHAO Menu commands from
the collar, and uses link messages to send appropriate commands to the
AO, depending on the detected collar type. It is not a complete
replacement of the Sub AO as it cannot allow as complete access to the
AO as that does, but it should provide most of the functionality while
being very easy to add to user's pre-existing AOs.

When the AO is switched on, but the collar sends a command to switch it
off, the script will take controls and detect for movement. While the
avatar is trying to move, the AO will told to switch the AO back on
again until movement ceases.

An alternative to this method would be to capture the AO's walk
animation and play the animation itself as a walk overrider. The
disadvantage of that is trying to capture the walk animation isn't
straightforward, particular when there is more than one walk configured.
The disadvantage of this method is that it does a lot more AO
triggering.