necromance some legacy concepts on taking controls
This is more or less Medea's legacy method to make it possible that
walks are overridden when AOs are "paused" so to say.

That means when the collar plays a pose, the AO is turned off, but the
script understands it as a "pause" and watches for input from the user
on arrow and WASD keys.

Once those are pressed, the AO comes back to life and correctly
overrides the walk rather than having the avatar slide on her knees.

There are some reasons to not to do this but taking into consideration
that this utility already goes into ZHAOII based AOs and that it's
almost three years later and we didn't get any fancier means/tools in
the LSL language to tackle that in more nerd-cool ways, I'm strongly
convinced that this is significantly more user friendly than AntiSlide
because unlike with client-side AOs, we don't need AntiSlide in this
case and can lift some complexity off the user's shoulder by making
things "just work".

#984