add basic command to check version + ...
... basic "rm" command to remove the script

Versioning here is really simple:

If the functionality changes we increment the first number, if we add
new compatibility or fix stuff we increment the decimal and if we
reached 10 decimals, we increment the whole/first number.

Keep this simple for users, as in decimal "something is fixed/added"
first number "ok, something important has changed".

The version I put is in accordance with that system.

#984